Move all differs into a separate 'differ' sub-package.
