+++ metadata-io/src/main/java/com/linkedin/metadata/timeline/ebean/EbeanTimelineService.java
@@ -13,6 +13,11 @@ import com.linkedin.metadata.timeline.data.ChangeCategory;
 import com.linkedin.metadata.timeline.data.ChangeEvent;
 import com.linkedin.metadata.timeline.data.ChangeTransaction;
 import com.linkedin.metadata.timeline.data.SemanticChangeType;
+import com.linkedin.metadata.timeline.ebean.differ.BasicDiffer;
+import com.linkedin.metadata.timeline.ebean.differ.DiffFactory;
+import com.linkedin.metadata.timeline.ebean.differ.Differ;
+import com.linkedin.metadata.timeline.ebean.differ.OwnershipDiffer;
+import com.linkedin.metadata.timeline.ebean.differ.SchemaDiffer;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Collection;
