+++ metadata-io/src/main/java/com/linkedin/metadata/timeline/ebean/differ/Differ.java
@@ -1,4 +1,4 @@
-package com.linkedin.metadata.timeline.ebean;
+package com.linkedin.metadata.timeline.ebean.differ;
 
 import com.github.fge.jsonpatch.JsonPatch;
 import com.linkedin.metadata.entity.ebean.EbeanAspectV2;
